move convertSecsToHumanReadable to datetime_utils
