+++ airflow/www/static/js/main.js
@@ -61,46 +61,13 @@ function changeDisplayedTimezone(tz) {
 
 const el = document.createElement('span');
 
+// eslint-disable-next-line import/prefer-default-export
 export function escapeHtml(text) {
   el.textContent = text;
   return el.innerHTML;
 }
 
 window.escapeHtml = escapeHtml;
-
-export function convertSecsToHumanReadable(seconds) {
-  const oriSeconds = seconds;
-  const floatingPart = oriSeconds - Math.floor(oriSeconds);
-
-  seconds = Math.floor(seconds);
-
-  const secondsPerHour = 60 * 60;
-  const secondsPerMinute = 60;
-
-  const hours = Math.floor(seconds / secondsPerHour);
-  seconds -= hours * secondsPerHour;
-
-  const minutes = Math.floor(seconds / secondsPerMinute);
-  seconds -= minutes * secondsPerMinute;
-
-  let readableFormat = '';
-  if (hours > 0) {
-    readableFormat += `${hours}Hours `;
-  }
-  if (minutes > 0) {
-    readableFormat += `${minutes}Min `;
-  }
-  if (seconds + floatingPart > 0) {
-    if (Math.floor(oriSeconds) === oriSeconds) {
-      readableFormat += `${seconds}Sec`;
-    } else {
-      seconds += floatingPart;
-      readableFormat += `${seconds.toFixed(3)}Sec`;
-    }
-  }
-  return readableFormat;
-}
-window.convertSecsToHumanReadable = convertSecsToHumanReadable;
 
 function postAsForm(url, parameters) {
   const form = $('<form></form>');
